lf iterate: lfops: split commands.py into per-command modules
Extract init, pr, land, commit, rebase, and summarize into separate files under lfops/. Shared git helpers moved to _helpers.py. Lazy imports in __init__.py to avoid circular dependencies.
