+++ src/loopflow/lf/__init__.py
@@ -9,7 +9,6 @@ import yaml
 
 from loopflow.lf.config import ConfigError, load_config
 from loopflow.lf.context import find_worktree_root, gather_task, list_all_tasks, _get_builtin_task
-from loopflow.init_check import check_init_status
 from loopflow.lf.pipelines import load_pipeline
