+++ src/loopflow/lfops/__init__.py
@@ -1,5 +1,16 @@
 """lfops: Loopflow operations CLI."""
 
-from loopflow.lfops.commands import app, main
 
-__all__ = ["app", "main"]
+def main() -> None:
+    """Entry point for lfops command."""
+    from loopflow.lfops.commands import main as _main
+    _main()
+
+
+def get_app():
+    """Get the Typer app (lazy import to avoid circular imports)."""
+    from loopflow.lfops.commands import app
+    return app
+
+
+__all__ = ["main", "get_app"]
